unit tests from engine.rs moved to tests/unit/engine.rs
